Separando componentes da tela de Home
